# cmp_backend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Consent Management Platform",
    "description": "\nConsent Management Platform.\n\nConventions that apply to every endpoint:\n\n* **Identifiers are uuids.** Integer primary keys never appear in a URL, a\n  response body, an export or a log. The one exception is `/c/{token}`, which\n  carries a capability rather than a reference.\n* **Lists are cursor-paginated.** `?limit=&cursor=&sort=`. Offset pagination\n  skips or repeats rows when the underlying set changes between pages.\n* **Unknown filters are rejected**, never ignored.\n* **403 means visible but not permitted.** Anything outside your scope is 404 -\n  a 403 would confirm it exists.\n",
    "contact": {
      "name": "Privacy Office",
      "email": "privacy@example.org"
    },
    "version": "0.1.0"
  },
  "paths": {
...
    "/purposes": {
      "get": {
        "tags": [
          "registry"
        ],
        "summary": "List Purposes",
        "operationId": "list_purposes_purposes_get",
        "parameters": [
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Status"
            }
          },
          {
            "name": "lawful_basis",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Lawful Basis"
            }
          },
          {
            "name": "q",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "maxLength": 100
                },
                {
                  "type": "null"
                }
              ],
              "title": "Q"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer",
                  "maximum": 200,
                  "minimum": 1
                },
                {
                  "type": "null"
                }
              ],
              "title": "Limit"
            }
          },
          {
            "name": "cursor",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "maxLength": 512
                },
                {
                  "type": "null"
                }
              ],
              "title": "Cursor"
            }
          },
          {
            "name": "sort",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "maxLength": 64
                },
                {
                  "type": "null"
                }
              ],
              "title": "Sort"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Page_PurposeOut_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "registry"
        ],
        "summary": "Create Purpose",
        "operationId": "create_purpose_purposes_post",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PurposeIn"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PurposeOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/purposes/{purpose_uuid}": {
      "get": {
        "tags": [
          "registry"
        ],
        "summary": "Get Purpose",
        "operationId": "get_purpose_purposes__purpose_uuid__get",
        "parameters": [
          {
            "name": "purpose_uuid",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Purpose Uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PurposeOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "registry"
        ],
        "summary": "Draft only",
        "operationId": "update_purpose_purposes__purpose_uuid__put",
        "parameters": [
          {
            "name": "purpose_uuid",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Purpose Uuid"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PurposeUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PurposeOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/purposes/{purpose_uuid}/activate": {
      "post": {
        "tags": [
          "registry"
        ],
        "summary": "Activate Purpose",
        "operationId": "activate_purpose_purposes__purpose_uuid__activate_post",
        "parameters": [
          {
            "name": "purpose_uuid",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Purpose Uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Acknowledged"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/purposes/{purpose_uuid}/retire": {
      "post": {
        "tags": [
          "registry"
        ],
        "summary": "Retire Purpose",
        "description": "Blocked while the purpose is attached to a published notice.\n\nRetiring it would leave a live notice offering a purpose the registry says\nno longer exists, and the consents already given against it unexplainable.",
        "operationId": "retire_purpose_purposes__purpose_uuid__retire_post",
        "parameters": [
          {
            "name": "purpose_uuid",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Purpose Uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Acknowledged"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/purposes/{purpose_uuid}/versions": {
      "get": {
        "tags": [
          "registry"
        ],
        "summary": "Purpose Versions",
        "operationId": "purpose_versions_purposes__purpose_uuid__versions_get",
        "parameters": [
          {
            "name": "purpose_uuid",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Purpose Uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/PurposeOut"
                  },
                  "title": "Response Purpose Versions Purposes  Purpose Uuid  Versions Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/purposes/{purpose_uuid}/usage": {
      "get": {
        "tags": [
          "registry"
        ],
        "summary": "Notices referencing this purpose",
        "description": "How the UI knows retirement is blocked before the user tries.",
        "operationId": "purpose_usage_purposes__purpose_uuid__usage_get",
        "parameters": [
          {
            "name": "purpose_uuid",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Purpose Uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Purpose Usage Purposes  Purpose Uuid  Usage Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Page_PurposeOut_": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/PurposeOut"
            },
            "type": "array",
            "title": "Items"
          },
          "next_cursor": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Next Cursor",
            "description": "Opaque. Pass back as ?cursor= for the next page."
          },
          "total": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Total",
            "description": "Total matching rows where it is cheap to know. Null means not counted."
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "items"
        ],
        "title": "Page[PurposeOut]"
      },
...
    }
  }
}